+++ src/module-info.java
@@ -1,4 +1,5 @@
 
 module QemuJuicy {
 	requires java.desktop;
+	requires java.datatransfer;
 }