+++ src/qemujuicy/vm/VMProperties.java
@@ -50,6 +50,7 @@ public class VMProperties extends Properties {
 	public static final String NETWORK = "network"; 
 	public static final String OS = "os"; 
 	public static final String QEMU_BOOT_MENU = "qemu.boot.menu"; 	
+	public static final String SOUND = "sound"; 
 	public static final String VERBOSE = "verbose";					// verbose output on stdout
 	public static final String VM_FILENAME = "vm.filename"; 		// the xml file containing this properties
 	public static final String VM_MEMORY_MB = "vm.memory.MB"; 		// in MB
@@ -108,6 +109,7 @@ public class VMProperties extends Properties {
 		checkProperty(NETWORK, "");
 		checkProperty(OS, "");
 		checkProperty(QEMU_BOOT_MENU, "");
+		checkProperty(SOUND, "");
 		checkProperty(VERBOSE, VMProperties.VERBOSE);
 		checkProperty(VM_FILENAME, "");
 		checkProperty(VM_MEMORY_MB, "");