+++ CHANGELOG.md
@@ -12,7 +12,7 @@ Details to start with: [README.md](https://github.com/openworld42/QemuJuicy/blob
 
 added: sound method/card (ComboBox + QEMU call)
 
-QemuJuicy help (Menu -> Help)
+QemuJuicy help (Menu -> Help, Button)
 
 added: copy QEMU call/parameters to clipboard (button + menu)